Upgrade NUnit to version 3 and upgrade to .net 4.5
Use Assert.Throws for tests that used the ExpectedException attribute, as it has been removed.
Use OneTimeSetUp and OneTimeTearDown instead of TestFixtureSetUp and TestFixtureTearDown attributes as they have been depracated.
Add description to Ignore attributes as a description is now required.
Update tests that use the file system to use the assembly path rather than the relative path.
Change tests to quit the process when shutting down the node.
Changed Assert.IsNullOrEmpty and Assert.IsNotNullOrEmpty to Assert.That as they have been removed.