Fix issue where duplicate entries could be written
Currently `NoStream` will be returned on exhausting the hash collision
read limit when attempting to search for the stream in question.
If `NoStream` is returned, it could cause duplicate entries to be written.

Instead of returning `NoStream`, it will now return
`EventNumber.Invalid` which is a little more accurate than `NoStream`

The IndexReader for ReadStreamEventsForward and
ReadStreamEventsBackward will now return `NoStream` which will result in the
status for clients being `StreamNotFound` upon getting
`EventNumber.Invalid` back for the last event number for a stream.

The IndexReader for ReadEvent will now return `NoStream` upon getting
`EventNumber.Invalid` back for the last event number for a stream.

IndexWriter will now throw a `WrongExpectedVersion` on attempting to
write to a stream where `EventNumber.Invalid` was returned for the last
event number for a stream.

Add index bitness version to vnode settings as well as to the
VNodeBuilder so that it can be set on the MiniNode for tests.